Split portal.js into routes/portal/ sub-routers (Chunk 9, M1)
Refactor the 1,798-line routes/portal.js into a sub-router directory
following the finance/, backup/, and members/ precedent. Pure move, no
behaviour change:

- profile.js  — /home dashboard, personal-details GET/PATCH, photo
  upload/delete/get, change-password, request-card
- groups.js   — browse plus join/leave
- calendar.js — event list and PDF download
- renewals.js — renewal-info, renew, renewal-confirm
- helpers.js  — date formatters and the stubbed email helpers
- index.js    — owns portal-auth middleware + feature gate, mounts subs

public.js updated to import portal/index.js. Backend suite green
(593 tests), lint and Prettier clean.
